refactor class/repo structure a bit
